move out upload related code into a mod
